Address review comments and reneable Metadata constructors since they work on Xcode 10 beta 6.
