+++ src/Network/NWProtocolMetadata.cs
@@ -38,7 +38,6 @@ namespace Network {
 	[TV (12,0), Mac (10,14), iOS (12,0)]
 	public class NWProtocolMetadata : NativeObject {
 
-#if false
 		// Officially listed on header files, but seems to not work on Mac/iOS
 		// https://bugreport.apple.com/web/?problemID=42443077
 		[DllImport (Constants.NetworkLibrary)]
@@ -55,7 +54,6 @@ namespace Network {
 		{
 			return new NWProtocolMetadata (nw_udp_create_metadata (), owns: true);
 		}
-#endif
 
 		public NWProtocolMetadata (IntPtr handle, bool owns) : base (handle, owns) {}
