+++ tests/xtro-sharpie/common-Network.ignore
@@ -1,4 +1,4 @@
-# the following protocols ae binded manualy as NativeObject
+# the following are binded manualy as NativeObject
 
 !missing-protocol! OS_nw_advertise_descriptor not bound
 !missing-protocol! OS_nw_connection not bound
@@ -15,10 +15,6 @@
 !missing-protocol! OS_nw_protocol_options not bound
 !missing-protocol! OS_nw_protocol_stack not bound
 
-# binded but with a name alias changes and xtro gets confused.
-!missing-pinvoke! nw_ip_create_metadata is not bound
-!missing-pinvoke! nw_udp_create_metadata is not bound
-
 # binded via a Marshal.ReadIntPtr (Dlfcn.dlsym ..
 !missing-field! _nw_parameters_configure_protocol_default_configuration not bound
 !missing-field! _nw_parameters_configure_protocol_disable not bound
@@ -26,3 +22,7 @@
 !missing-field! _nw_content_context_default_stream not bound
 !missing-field! _nw_content_context_final_send not bound
 !missing-field! _nw_connection_send_idempotent_content not bound
+
+# no need, please see comment: https://github.com/xamarin/xamarin-macios/pull/4251#issuecomment-410815837
+!missing-pinvoke! nw_release is not bound
+!missing-pinvoke! nw_retain is not bound
