+++ tests/xtro-sharpie/macOS-Network.todo
@@ -1,6 +1,4 @@
 !missing-pinvoke! nw_endpoint_get_address is not bound
 !missing-pinvoke! nw_path_monitor_set_cancel_handler is not bound
 !missing-pinvoke! nw_path_monitor_set_update_handler is not bound
-!missing-pinvoke! nw_release is not bound
-!missing-pinvoke! nw_retain is not bound
 !missing-protocol! OS_nw_object not bound
